refactor: change `SearchBox` to a class comp
Change `SearchBox` to a class comp to maintain its own state
and handlers. `searchQuery` or `query` is no longer needed
in the App's state since other components take query from route.

Avail router params and history from `this.props`. `useHistory` and
`useParams` hooks need not and should not be use in class components.