fix: add rules to rewrite all routes to index page
Problem:
Visiting the routes such as '/browse' directly on local site (localhost)
works fine. But on the remote site, this causes a Page Not Found
(404) error.

Reason:
URL is intrepreted in server-side first and then in client-side.
So, when the routes are visited directly,
react-router is not accessible and server doesn't know to handle routes.

Fix:
Rewrite all routes to index page where 'index.html' will be served.
Then, react-router will load
and  handle the routes.

More info: [Rewrites and proxies](https://kutt.it/yhfgj1) -- Netlify